Add the 'std' feature
Finally lets combine compile in no_std environments!
