+++ src/primitives.rs
@@ -1,15 +1,18 @@
-use std::any::Any;
-use std::cell::UnsafeCell;
-use std::collections::VecDeque;
-use std::error::Error as StdError;
-use std::fmt;
+use lib::fmt;
+
+#[cfg(feature = "std")]
 use std::io::{Bytes, Read};
+#[cfg(feature = "std")]
+use std::error::Error as StdError;
 
 use self::FastResult::*;
 
 use ErrorOffset;
 use combinator::{and_then, expected, flat_map, map, message, or, skip, then, with, AndThen,
                  Expected, FlatMap, Iter, Map, Message, Or, Skip, Then, With};
+
+#[cfg(feature = "std")]
+use simple::Error;
 
 #[macro_export]
 macro_rules! ctry {
@@ -59,105 +62,6 @@ impl PointerOffset {
     }
 }
 
-/// Enum holding error information. Variants are defined for `Stream::Item` and `Stream::Range` as
-/// well as string variants holding simple descriptions.
-///
-/// As there is implementations of `From` for `String` and `&'static str` the
-/// constructor need not be used directly as calling `msg.into()` should turn a message into the
-/// correct `Info` variant.
-#[derive(Clone, Debug)]
-pub enum Info<T, R> {
-    Token(T),
-    Range(R),
-    Owned(String),
-    Borrowed(&'static str),
-}
-
-impl<T, R> From<SimpleInfo<T, R>> for Info<T, R> {
-    fn from(info: SimpleInfo<T, R>) -> Self {
-        match info {
-            SimpleInfo::Token(b) => Info::Token(b),
-            SimpleInfo::Range(b) => Info::Range(b),
-            SimpleInfo::Borrowed(b) => Info::Borrowed(b),
-        }
-    }
-}
-
-impl<T, R> Info<T, R> {
-    pub fn map_token<F, U>(self, f: F) -> Info<U, R>
-    where
-        F: FnOnce(T) -> U,
-    {
-        use self::Info::*;
-        match self {
-            Token(t) => Token(f(t)),
-            Range(r) => Range(r),
-            Owned(s) => Owned(s),
-            Borrowed(x) => Borrowed(x),
-        }
-    }
-
-    pub fn map_range<F, S>(self, f: F) -> Info<T, S>
-    where
-        F: FnOnce(R) -> S,
-    {
-        use self::Info::*;
-        match self {
-            Token(t) => Token(t),
-            Range(r) => Range(f(r)),
-            Owned(s) => Owned(s),
-            Borrowed(x) => Borrowed(x),
-        }
-    }
-}
-
-impl<T: PartialEq, R: PartialEq> PartialEq for Info<T, R> {
-    fn eq(&self, other: &Info<T, R>) -> bool {
-        match (self, other) {
-            (&Info::Token(ref l), &Info::Token(ref r)) => l == r,
-            (&Info::Range(ref l), &Info::Range(ref r)) => l == r,
-            (&Info::Owned(ref l), &Info::Owned(ref r)) => l == r,
-            (&Info::Borrowed(l), &Info::Owned(ref r)) => l == r,
-            (&Info::Owned(ref l), &Info::Borrowed(r)) => l == r,
-            (&Info::Borrowed(l), &Info::Borrowed(r)) => l == r,
-            _ => false,
-        }
-    }
-}
-impl<T: fmt::Display, R: fmt::Display> fmt::Display for Info<T, R> {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match *self {
-            Info::Token(ref c) => write!(f, "{}", c),
-            Info::Range(ref c) => write!(f, "{}", c),
-            Info::Owned(ref s) => write!(f, "{}", s),
-            Info::Borrowed(s) => write!(f, "{}", s),
-        }
-    }
-}
-
-impl<R> From<char> for Info<char, R> {
-    fn from(s: char) -> Info<char, R> {
-        Info::Token(s)
-    }
-}
-impl<T, R> From<String> for Info<T, R> {
-    fn from(s: String) -> Info<T, R> {
-        Info::Owned(s)
-    }
-}
-
-impl<T, R> From<&'static str> for Info<T, R> {
-    fn from(s: &'static str) -> Info<T, R> {
-        Info::Borrowed(s)
-    }
-}
-
-impl<R> From<u8> for Info<u8, R> {
-    fn from(s: u8) -> Info<u8, R> {
-        Info::Token(s)
-    }
-}
-
 impl<R> From<char> for SimpleInfo<char, R> {
     fn from(s: char) -> SimpleInfo<char, R> {
         SimpleInfo::Token(s)
@@ -173,366 +77,6 @@ impl<T, R> From<&'static str> for SimpleInfo<T, R> {
 impl<R> From<u8> for SimpleInfo<u8, R> {
     fn from(s: u8) -> SimpleInfo<u8, R> {
         SimpleInfo::Token(s)
-    }
-}
-
-/// Enum used to store information about an error that has occurred during parsing.
-#[derive(Debug)]
-pub enum Error<T, R> {
-    /// Error indicating an unexpected token has been encountered in the stream
-    Unexpected(Info<T, R>),
-    /// Error indicating that the parser expected something else
-    Expected(Info<T, R>),
-    /// Generic message
-    Message(Info<T, R>),
-    /// Variant for containing other types of errors
-    Other(Box<StdError + Send + Sync>),
-}
-
-impl<T, R> From<SimpleError<T, R>> for Error<T, R> {
-    fn from(err: SimpleError<T, R>) -> Self {
-        match err {
-            SimpleError::Message(info) => Error::Message(info.into()),
-            SimpleError::Expected(info) => Error::Expected(info.into()),
-            SimpleError::Unexpected(info) => Error::Unexpected(info.into()),
-        }
-    }
-}
-
-impl<Item, Range> StreamingError<Item, Range> for Error<Item, Range>
-where
-    Item: PartialEq,
-    Range: PartialEq,
-{
-    #[inline]
-    fn unexpected_token(token: Item) -> Self {
-        Error::Unexpected(Info::Token(token))
-    }
-    #[inline]
-    fn unexpected_range(token: Range) -> Self {
-        Error::Unexpected(Info::Range(token))
-    }
-    #[inline]
-    fn unexpected_message<T>(msg: T) -> Self
-    where
-        T: fmt::Display,
-    {
-        Error::Unexpected(Info::Owned(msg.to_string()))
-    }
-    #[inline]
-    fn unexpected_static_message(msg: &'static str) -> Self {
-        Error::Unexpected(Info::Borrowed(msg))
-    }
-
-    #[inline]
-    fn expected_token(token: Item) -> Self {
-        Error::Expected(Info::Token(token))
-    }
-    #[inline]
-    fn expected_range(token: Range) -> Self {
-        Error::Expected(Info::Range(token))
-    }
-    #[inline]
-    fn expected_message<T>(msg: T) -> Self
-    where
-        T: fmt::Display,
-    {
-        Error::Expected(Info::Owned(msg.to_string()))
-    }
-    #[inline]
-    fn expected_static_message(msg: &'static str) -> Self {
-        Error::Expected(Info::Borrowed(msg))
-    }
-
-    #[inline]
-    fn message_message<T>(msg: T) -> Self
-    where
-        T: fmt::Display,
-    {
-        Error::Message(Info::Owned(msg.to_string()))
-    }
-    #[inline]
-    fn message_static_message(msg: &'static str) -> Self {
-        Error::Message(Info::Borrowed(msg))
-    }
-    #[inline]
-    fn message_token(token: Item) -> Self {
-        Error::Message(Info::Token(token))
-    }
-    #[inline]
-    fn message_range(token: Range) -> Self {
-        Error::Message(Info::Range(token))
-    }
-
-    #[inline]
-    fn other<E>(err: E) -> Self
-    where
-        E: StdError + Send + Sync + 'static,
-    {
-        err.into()
-    }
-
-
-    #[inline]
-    fn into_other<T>(self) -> T
-    where
-        T: StreamingError<Item, Range>,
-    {
-        match self {
-            Error::Unexpected(info) => match info {
-                Info::Token(x) => T::unexpected_token(x),
-                Info::Range(x) => T::unexpected_range(x),
-                Info::Borrowed(x) => T::unexpected_static_message(x),
-                Info::Owned(x) => T::unexpected_message(x),
-            },
-            Error::Expected(info) => match info {
-                Info::Token(x) => T::expected_token(x),
-                Info::Range(x) => T::expected_range(x),
-                Info::Borrowed(x) => T::expected_static_message(x),
-                Info::Owned(x) => T::expected_message(x),
-            },
-            Error::Message(info) => match info {
-                Info::Token(x) => T::expected_token(x),
-                Info::Range(x) => T::expected_range(x),
-                Info::Borrowed(x) => T::expected_static_message(x),
-                Info::Owned(x) => T::expected_message(x),
-            },
-            Error::Other(err) => T::message_message(err),
-        }
-    }
-}
-
-impl<Item, Range, Position> ParsingError<Item, Range, Position> for Error<Item, Range>
-where
-    Item: PartialEq,
-    Range: PartialEq,
-    Position: Default,
-{
-    type StreamError = Self;
-    #[inline]
-    fn empty(_: Position) -> Self {
-        Self::message_static_message("")
-    }
-    #[inline]
-    fn from_error(_: Position, err: Self::StreamError) -> Self {
-        err
-    }
-
-    #[inline]
-    fn set_position(&mut self, _position: Position) {}
-
-
-    #[inline]
-    fn add(&mut self, err: Self::StreamError) {
-        *self = err;
-    }
-
-    #[inline]
-    fn set_expected<F>(self_: &mut Tracked<Self>, info: Self::StreamError, f: F)
-    where
-        F: FnOnce(&mut Tracked<Self>),
-    {
-        f(self_);
-        self_.error = info;
-    }
-
-    #[inline]
-    fn into_other<T>(self) -> T
-    where
-        T: ParsingError<Item, Range, Position>,
-    {
-        T::from_error(Position::default(), StreamingError::into_other(self))
-    }
-}
-
-impl<Item, Range, Position> ParsingError<Item, Range, Position>
-    for ParseError<Position, Item, Range>
-where
-    Position: Default,
-    Item: PartialEq,
-    Range: PartialEq,
-    Position: Ord,
-{
-    type StreamError = Error<Item, Range>;
-    #[inline]
-    fn empty(pos: Position) -> Self {
-        ParseError::empty(pos)
-    }
-    #[inline]
-    fn from_error(position: Position, err: Self::StreamError) -> Self {
-        ParseError::new(position, Error::from(err))
-    }
-
-    #[inline]
-    fn set_position(&mut self, position: Position) {
-        self.position = position;
-    }
-
-    #[inline]
-    fn merge(self, other: Self) -> Self {
-        ParseError::merge(self, other)
-    }
-
-    #[inline]
-    fn add(&mut self, err: Self::StreamError) {
-        self.add_error(err);
-    }
-
-    #[inline]
-    fn set_expected<F>(self_: &mut Tracked<Self>, info: Self::StreamError, f: F)
-    where
-        F: FnOnce(&mut Tracked<Self>),
-    {
-        let start = self_.error.errors.len();
-        f(self_);
-        // Replace all expected errors that were added from the previous add_error
-        // with this expected error
-        let mut i = 0;
-        self_.error.errors.retain(|e| if i < start {
-            i += 1;
-            true
-        } else {
-            match *e {
-                Error::Expected(_) => false,
-                _ => true,
-            }
-        });
-        self_.error.add(info);
-    }
-
-    #[inline]
-    fn into_other<T>(mut self) -> T
-    where
-        T: ParsingError<Item, Range, Position>,
-    {
-        match self.errors.pop() {
-            Some(err) => T::from_error(self.position, StreamingError::into_other(err)),
-            None => T::empty(self.position),
-        }
-    }
-}
-
-impl<T, R> Error<T, R> {
-    pub fn map_token<F, U>(self, f: F) -> Error<U, R>
-    where
-        F: FnOnce(T) -> U,
-    {
-        use self::Error::*;
-        match self {
-            Unexpected(x) => Unexpected(x.map_token(f)),
-            Expected(x) => Expected(x.map_token(f)),
-            Message(x) => Message(x.map_token(f)),
-            Other(x) => Other(x),
-        }
-    }
-
-    pub fn map_range<F, S>(self, f: F) -> Error<T, S>
-    where
-        F: FnOnce(R) -> S,
-    {
-        use self::Error::*;
-        match self {
-            Unexpected(x) => Unexpected(x.map_range(f)),
-            Expected(x) => Expected(x.map_range(f)),
-            Message(x) => Message(x.map_range(f)),
-            Other(x) => Other(x),
-        }
-    }
-}
-
-impl<T: PartialEq, R: PartialEq> PartialEq for Error<T, R> {
-    fn eq(&self, other: &Error<T, R>) -> bool {
-        match (self, other) {
-            (&Error::Unexpected(ref l), &Error::Unexpected(ref r)) |
-            (&Error::Expected(ref l), &Error::Expected(ref r)) |
-            (&Error::Message(ref l), &Error::Message(ref r)) => l == r,
-            _ => false,
-        }
-    }
-}
-
-impl<E, T, R> From<E> for Error<T, R>
-where
-    E: StdError + 'static + Send + Sync,
-{
-    fn from(e: E) -> Error<T, R> {
-        Error::Other(Box::new(e))
-    }
-}
-
-impl<T, R> Error<T, R> {
-    /// Returns the `end_of_input` error.
-    pub fn end_of_input() -> Error<T, R> {
-        Error::Unexpected("end of input".into())
-    }
-
-    /// Formats a slice of errors in a human readable way.
-    ///
-    /// ```rust
-    /// # extern crate combine;
-    /// # use combine::*;
-    /// # use combine::char::*;
-    ///
-    /// # fn main() {
-    /// let input = r"
-    ///   ,123
-    /// ";
-    /// let result = spaces().with(char('.').or(char('a')).or(digit()))
-    ///     .simple_parse(State::new(input));
-    /// let m = format!("{}", result.unwrap_err());
-    /// let expected = r"Parse error at line: 2, column: 3
-    /// Unexpected `,`
-    /// Expected `.`, `a` or `digit`
-    /// ";
-    /// assert_eq!(m, expected);
-    /// # }
-    /// ```
-    pub fn fmt_errors(errors: &[Error<T, R>], f: &mut fmt::Formatter) -> fmt::Result
-    where
-        T: fmt::Display,
-        R: fmt::Display,
-    {
-        // First print the token that we did not expect
-        // There should really just be one unexpected message at this point though we print them
-        // all to be safe
-        let unexpected = errors.iter().filter(|e| match **e {
-            Error::Unexpected(_) => true,
-            _ => false,
-        });
-        for error in unexpected {
-            try!(writeln!(f, "{}", error));
-        }
-
-        // Then we print out all the things that were expected in a comma separated list
-        // 'Expected 'a', 'expression' or 'let'
-        let iter = || {
-            errors.iter().filter_map(|e| match *e {
-                Error::Expected(ref err) => Some(err),
-                _ => None,
-            })
-        };
-        let expected_count = iter().count();
-        for (i, message) in iter().enumerate() {
-            let s = match i {
-                0 => "Expected",
-                _ if i < expected_count - 1 => ",",
-                // Last expected message to be written
-                _ => " or",
-            };
-            try!(write!(f, "{} `{}`", s, message));
-        }
-        if expected_count != 0 {
-            try!(writeln!(f, ""));
-        }
-        // If there are any generic messages we print them out last
-        let messages = errors.iter().filter(|e| match **e {
-            Error::Message(_) | Error::Other(_) => true,
-            _ => false,
-        });
-        for error in messages {
-            try!(writeln!(f, "{}", error));
-        }
-        Ok(())
     }
 }
 
@@ -682,184 +226,6 @@ impl<T> Consumed<T> {
     }
 }
 
-/// Alias over `ParseError` for `StreamOnce` types
-pub type StreamError<S> = ParseError<
-    <S as StreamOnce>::Position,
-    <S as StreamOnce>::Item,
-    <S as StreamOnce>::Range,
->;
-
-/// Struct which hold information about an error that occured at a specific position.
-/// Can hold multiple instances of `Error` if more that one error occured in the same position.
-#[derive(Debug, PartialEq)]
-pub struct ParseError<P, I, R> {
-    /// The position where the error occurred
-    pub position: P,
-    /// A vector containing specific information on what errors occurred at `position`. Usually
-    /// a fully formed message contains one `Unexpected` error and one or more `Expected` errors.
-    /// `Message` and `Other` may also appear (`combine` never generates these errors on its own)
-    /// and may warrant custom handling.
-    pub errors: Vec<Error<I, R>>,
-}
-
-impl<P, I, R> ParseError<P, I, R> {
-    /// Constructs a new `ParseError` which occurred at `position`.
-    #[inline]
-    pub fn new(position: P, error: Error<I, R>) -> ParseError<P, I, R> {
-        ParseError::from_errors(position, vec![error])
-    }
-
-    /// Constructs an error with no other information than the position it occurred at.
-    #[inline]
-    pub fn empty(position: P) -> ParseError<P, I, R> {
-        ParseError::from_errors(position, vec![])
-    }
-
-    /// Constructs a `ParseError` with multiple causes.
-    #[inline]
-    pub fn from_errors(position: P, errors: Vec<Error<I, R>>) -> ParseError<P, I, R> {
-        ParseError {
-            position: position,
-            errors: errors,
-        }
-    }
-
-    /// Constructs an end of input error. Should be returned by parsers which encounter end of
-    /// input unexpectedly.
-    #[inline]
-    pub fn end_of_input(position: P) -> ParseError<P, I, R> {
-        ParseError::new(position, Error::end_of_input())
-    }
-
-    /// Adds an error if `error` does not exist in this `ParseError` already (as determined byte
-    /// `PartialEq`).
-    pub fn add_error(&mut self, error: Error<I, R>)
-    where
-        I: PartialEq,
-        R: PartialEq,
-    {
-        // Don't add duplicate errors
-        if self.errors.iter().all(|err| *err != error) {
-            self.errors.push(error);
-        }
-    }
-
-    /// Remvoes all `Expected` errors in `self` and adds `info` instead.
-    pub fn set_expected(&mut self, info: Info<I, R>) {
-        // Remove all other expected messages
-        self.errors.retain(|e| match *e {
-            Error::Expected(_) => false,
-            _ => true,
-        });
-        self.errors.push(Error::Expected(info));
-    }
-
-    /// Merges two `ParseError`s. If they exist at the same position the errors of `other` are
-    /// added to `self` (using `add_error` to skip duplicates). If they are not at the same
-    /// position the error furthest ahead are returned, ignoring the other `ParseError`.
-    pub fn merge(mut self, mut other: ParseError<P, I, R>) -> ParseError<P, I, R>
-    where
-        P: Ord,
-        I: PartialEq,
-        R: PartialEq,
-    {
-        use std::cmp::Ordering;
-        // Only keep the errors which occurred after consuming the most amount of data
-        match self.position.cmp(&other.position) {
-            Ordering::Less => other,
-            Ordering::Greater => self,
-            Ordering::Equal => {
-                for message in other.errors.drain(..) {
-                    self.add_error(message);
-                }
-                self
-            }
-        }
-    }
-
-    /// Maps the position to a new value
-    pub fn map_position<F, Q>(self, f: F) -> ParseError<Q, I, R>
-    where
-        F: FnOnce(P) -> Q,
-    {
-        ParseError::from_errors(f(self.position), self.errors)
-    }
-
-    /// Maps all token variants to a new value
-    pub fn map_token<F, U>(self, mut f: F) -> ParseError<P, U, R>
-    where
-        F: FnMut(I) -> U,
-    {
-        ParseError::from_errors(
-            self.position,
-            self.errors
-                .into_iter()
-                .map(|error| error.map_token(&mut f))
-                .collect(),
-        )
-    }
-
-    /// Maps all range variants to a new value.
-    ///
-    /// ```
-    /// use combine::Parser;
-    /// use combine::range::range;
-    /// println!(
-    ///     "{}",
-    ///     range(&"HTTP"[..])
-    ///         .simple_parse("HTT")
-    ///         .unwrap_err()
-    ///         .map_range(|bytes| format!("{:?}", bytes))
-    /// );
-    /// ```
-    pub fn map_range<F, S>(self, mut f: F) -> ParseError<P, I, S>
-    where
-        F: FnMut(R) -> S,
-    {
-        ParseError::from_errors(
-            self.position,
-            self.errors
-                .into_iter()
-                .map(|error| error.map_range(&mut f))
-                .collect(),
-        )
-    }
-}
-
-impl<P, I, R> StdError for ParseError<P, I, R>
-where
-    P: fmt::Display + fmt::Debug + Any,
-    I: fmt::Display + fmt::Debug + Any,
-    R: fmt::Display + fmt::Debug + Any,
-{
-    fn description(&self) -> &str {
-        "parse error"
-    }
-}
-
-impl<P, I, R> fmt::Display for ParseError<P, I, R>
-where
-    P: fmt::Display,
-    I: fmt::Display,
-    R: fmt::Display,
-{
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        try!(writeln!(f, "Parse error at {}", self.position));
-        Error::fmt_errors(&self.errors, f)
-    }
-}
-
-impl<T: fmt::Display, R: fmt::Display> fmt::Display for Error<T, R> {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match *self {
-            Error::Unexpected(ref c) => write!(f, "Unexpected `{}`", c),
-            Error::Expected(ref s) => write!(f, "Expected `{}`", s),
-            Error::Message(ref msg) => msg.fmt(f),
-            Error::Other(ref err) => err.fmt(f),
-        }
-    }
-}
-
 /// A type alias over the specific `Result` type used by parsers to indicate wether they were
 /// successful or not.
 /// `O` is the type that is output on success.
@@ -933,6 +299,7 @@ pub trait StreamingError<Item, Range>: Sized + PartialEq {
         }
     }
 
+    #[cfg(feature = "std")]
     fn other<E>(err: E) -> Self
     where
         E: StdError + Send + Sync + 'static,
@@ -1147,66 +514,6 @@ where
     I: StreamOnce + Positioned + Clone,
     I::Error: ParsingError<I::Item, I::Range, I::Position>,
 {
-}
-
-#[derive(Copy, Clone, Debug)]
-pub struct SimpleStream<S>(pub S);
-
-impl<S> StreamOnce for SimpleStream<S>
-where
-    S: StreamOnce + Positioned,
-    S::Position: Default,
-{
-    type Item = S::Item;
-    type Range = S::Range;
-    type Position = S::Position;
-    type Error = StreamError<S>;
-
-    #[inline]
-    fn uncons<E>(&mut self) -> Result<Self::Item, E>
-    where
-        E: StreamingError<Self::Item, Self::Range>,
-    {
-        match self.0.uncons() {
-            Ok(t) => Ok(t),
-            Err(err) => Err(err),
-        }
-    }
-}
-
-impl<S> RangeStreamOnce for SimpleStream<S>
-where
-    S: RangeStream,
-    S::Error: ParsingError<S::Item, S::Range, S::Position>,
-    S::Position: Default,
-{
-    #[inline]
-    fn uncons_range(&mut self, size: usize) -> Result<Self::Range, Self::Error> {
-        self.0.uncons_range(size).map_err(S::Error::into_other)
-    }
-
-    #[inline]
-    fn uncons_while<F>(&mut self, f: F) -> Result<Self::Range, Self::Error>
-    where
-        F: FnMut(Self::Item) -> bool,
-    {
-        self.0.uncons_while(f).map_err(S::Error::into_other)
-    }
-
-    #[inline]
-    fn distance(&self, end: &Self) -> usize {
-        self.0.distance(&end.0)
-    }
-}
-
-impl<S> Positioned for SimpleStream<S>
-where
-    S: StreamOnce + Positioned,
-    S::Position: Default,
-{
-    fn position(&self) -> S::Position {
-        self.0.position()
-    }
 }
 
 #[inline]
@@ -1726,10 +1033,12 @@ where
     }
 }
 
+#[cfg(feature = "std")]
 pub struct ReadStream<R> {
     bytes: Bytes<R>,
 }
 
+#[cfg(feature = "std")]
 impl<R: Read> StreamOnce for ReadStream<R> {
     type Item = u8;
     type Range = u8;
@@ -1749,6 +1058,7 @@ impl<R: Read> StreamOnce for ReadStream<R> {
     }
 }
 
+#[cfg(feature = "std")]
 impl<R> ReadStream<R>
 where
     R: Read,
@@ -1759,10 +1069,12 @@ where
     ///     such as `BufferedStreamRef` and `State` to become a `Stream` which can be parsed
     ///
     /// ```rust
+    /// # #![cfg(feature = "std")]
     /// # extern crate combine;
     /// use combine::*;
     /// use combine::byte::*;
-    /// use combine::primitives::{BufferedStream, ReadStream};
+    /// use combine::primitives::ReadStream;
+    /// use combine::buffered_stream::BufferedStream;
     /// use combine::state::State;
     /// use std::io::Read;
     ///
@@ -1928,19 +1240,20 @@ pub trait Parser {
     /// [`ParseError`] otherwise.
     ///
     /// [`ParseError`]: struct.ParseError.html
-    fn simple_parse<I>(&mut self, input: I) -> Result<(Self::Output, I), StreamError<I>>
+    #[cfg(feature = "std")]
+    fn simple_parse<I>(&mut self, input: I) -> Result<(Self::Output, I), ::simple::StreamError<I>>
     where
         I: Stream,
-        SimpleStream<I>: StreamOnce<
+        ::simple::SimpleStream<I>: StreamOnce<
             Item = I::Item,
             Range = I::Range,
-            Error = StreamError<SimpleStream<I>>,
+            Error = ::simple::StreamError<::simple::SimpleStream<I>>,
             Position = I::Position,
         >,
         I::Position: Default,
-        Self: Sized + Parser<Input = SimpleStream<I>>,
+        Self: Sized + Parser<Input = ::simple::SimpleStream<I>>,
     {
-        match self.parse_stream(SimpleStream(input)) {
+        match self.parse_stream(::simple::SimpleStream(input)) {
             Ok((v, state)) => Ok((v, state.into_inner().0)),
             Err(error) => Err(error.into_inner().error),
         }
@@ -2178,10 +1491,12 @@ pub trait Parser {
     /// the rest of the input.
     ///
     /// ```
+    /// # #![cfg(feature = "std")]
     /// # extern crate combine;
     /// # use combine::*;
     /// # use combine::char::digit;
-    /// # use combine::primitives::{Consumed, Error, SimpleStream};
+    /// # use combine::primitives::Consumed;
+    /// # use combine::simple::{Error, SimpleStream};
     /// # fn main() {
     /// let result = digit()
     ///     .then(|d| parser(move |input| {
@@ -2257,9 +1572,10 @@ pub trait Parser {
     /// Parses with `self` and if it fails, adds the message `msg` to the error.
     ///
     /// ```
+    /// # #![cfg(feature = "std")]
     /// # extern crate combine;
     /// # use combine::*;
-    /// # use combine::primitives::Error;
+    /// # use combine::simple::Error;
     /// # use combine::state::SourcePosition;
     /// # fn main() {
     /// let result = token('9')
@@ -2287,9 +1603,10 @@ pub trait Parser {
     /// replaced by `msg`. `msg` is then used in error messages as "Expected `msg`".
     ///
     /// ```
+    /// # #![cfg(feature = "std")]
     /// # extern crate combine;
     /// # use combine::*;
-    /// # use combine::primitives::Error;
+    /// # use combine::simple::Error;
     /// # use combine::state::SourcePosition;
     /// # fn main() {
     /// let result = token('9')
@@ -2388,6 +1705,7 @@ pub trait Parser {
     /// assert_eq!(result, Ok((('a', 'c'), "")));
     /// # }
     /// ```
+    #[cfg(feature = "std")]
     fn boxed<'a>(self) -> Box<Parser<Input = Self::Input, Output = Self::Output> + 'a>
     where
         Self: Sized + 'a,
@@ -2424,6 +1742,8 @@ where
         (**self).add_error(error)
     }
 }
+
+#[cfg(feature = "std")]
 impl<I, O, P: ?Sized> Parser for Box<P>
 where
     I: Stream,
@@ -2453,197 +1773,6 @@ where
     }
 }
 
-/// A `BufferedStreamRef` wraps an instance `StreamOnce`, allowing it to be used as a `Stream`.
-pub struct BufferedStreamRef<'a, I>
-where
-    I: StreamOnce + Positioned + 'a,
-    I::Item: 'a,
-{
-    offset: usize,
-    buffer: &'a BufferedStream<I>,
-}
-
-impl<'a, I> fmt::Debug for BufferedStreamRef<'a, I>
-where
-    I: StreamOnce + Positioned + 'a,
-    I::Item: 'a,
-{
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let buffer_offset = unsafe { (*self.buffer.buffer.get()).offset };
-        write!(
-            f,
-            "BufferedStreamRef {{ offset: {:?} buffer_offset: {:?} }}",
-            self.offset,
-            buffer_offset
-        )
-    }
-}
-
-impl<'a, I> Clone for BufferedStreamRef<'a, I>
-where
-    I: StreamOnce + Positioned + 'a,
-    I::Position: Clone,
-    I::Item: 'a,
-{
-    fn clone(&self) -> BufferedStreamRef<'a, I> {
-        BufferedStreamRef {
-            offset: self.offset,
-            buffer: self.buffer,
-        }
-    }
-}
-
-pub struct BufferedStream<I>
-where
-    I: StreamOnce + Positioned,
-{
-    buffer: UnsafeCell<BufferedStreamInner<I>>,
-}
-
-struct BufferedStreamInner<I>
-where
-    I: StreamOnce + Positioned,
-{
-    offset: usize,
-    iter: I,
-    buffer: VecDeque<(I::Item, I::Position)>,
-}
-
-impl<I> BufferedStreamInner<I>
-where
-    I: StreamOnce + Positioned,
-    I::Position: Clone,
-    I::Item: Clone,
-{
-    #[inline]
-    fn uncons<E>(&mut self, offset: usize) -> Result<I::Item, E>
-    where
-        E: StreamingError<I::Item, I::Range>,
-    {
-        if offset >= self.offset {
-            let position = self.iter.position();
-            let item = try!(self.iter.uncons());
-            self.offset += 1;
-            // We want the VecDeque to only keep the last .capacity() elements so we need to remove
-            // an element if it gets to large
-            if self.buffer.len() == self.buffer.capacity() {
-                self.buffer.pop_front();
-            }
-            self.buffer.push_back((item.clone(), position.clone()));
-            Ok(item)
-        } else if offset < self.offset - self.buffer.len() {
-            // We have backtracked to far
-            Err(StreamingError::message_static_message(
-                "Backtracked to far".into(),
-            ))
-        } else {
-            Ok(
-                self.buffer[self.buffer.len() - (self.offset - offset)]
-                    .0
-                    .clone(),
-            )
-        }
-    }
-
-    #[inline]
-    fn position(&self, offset: usize) -> I::Position {
-        if offset >= self.offset {
-            self.iter.position()
-        } else if offset < self.offset - self.buffer.len() {
-            self.buffer
-                .front()
-                .expect("Atleast 1 element in the buffer")
-                .1
-                .clone()
-        } else {
-            self.buffer[self.buffer.len() - (self.offset - offset)]
-                .1
-                .clone()
-        }
-    }
-}
-
-impl<I> BufferedStream<I>
-where
-    I: StreamOnce + Positioned,
-    I::Position: Clone,
-    I::Item: Clone,
-{
-    /// Constructs a new `BufferedStream` froma a `StreamOnce` instance with a `lookahead`
-    /// number of elements that can be stored in the buffer.
-    ///
-    /// `BufferedStream` do not implement `Stream` itself. To retrieve a value which implement
-    /// `Stream`, `as_stream` must be called.
-    pub fn new(iter: I, lookahead: usize) -> BufferedStream<I> {
-        BufferedStream {
-            buffer: UnsafeCell::new(BufferedStreamInner {
-                offset: 0,
-                iter: iter,
-                buffer: VecDeque::with_capacity(lookahead),
-            }),
-        }
-    }
-
-    /// Creates a `BufferedStreamRef` which implements `Stream`.
-    ///
-    /// `BufferedStreamRef` always implement `Stream` allowing one-shot streams to used as if it
-    /// could be used multiple times.
-    pub fn as_stream(&self) -> BufferedStreamRef<I> {
-        BufferedStreamRef {
-            offset: 0,
-            buffer: self,
-        }
-    }
-    #[inline]
-    fn uncons<E>(&self, offset: usize) -> Result<I::Item, E>
-    where
-        E: StreamingError<I::Item, I::Range>,
-    {
-        unsafe { (*self.buffer.get()).uncons(offset) }
-    }
-    #[inline(always)]
-    fn position(&self, offset: usize) -> I::Position {
-        unsafe { (*self.buffer.get()).position(offset) }
-    }
-}
-
-impl<'a, I> BufferedStreamRef<'a, I>
-where
-    I: StreamOnce + Positioned,
-{
-}
-
-impl<'a, I> Positioned for BufferedStreamRef<'a, I>
-where
-    I: StreamOnce + Positioned + 'a,
-{
-    #[inline(always)]
-    fn position(&self) -> Self::Position {
-        self.buffer.position(self.offset)
-    }
-}
-
-impl<'a, I> StreamOnce for BufferedStreamRef<'a, I>
-where
-    I: StreamOnce + Positioned + 'a,
-    I::Item: Clone + PartialEq + 'a,
-{
-    type Item = I::Item;
-    type Range = I::Range;
-    type Position = I::Position;
-    type Error = I::Error;
-
-    #[inline]
-    fn uncons<E>(&mut self) -> Result<I::Item, E>
-    where
-        E: StreamingError<Self::Item, Self::Range>,
-    {
-        let value = try!(self.buffer.uncons(self.offset));
-        self.offset += 1;
-        Ok(value)
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -2656,6 +1785,11 @@ mod tests {
         let s: &[u8] = &[];
         assert_eq!(SliceStream(s).uncons_range(0), Ok(&[][..]));
     }
+}
+
+#[cfg(all(feature = "std", test))]
+mod tests_std {
+    use super::*;
 
     #[derive(Clone, PartialEq, Debug)]
     struct CloneOnly {
