+++ travis.sh
@@ -5,4 +5,5 @@ cargo build &&
     cargo test --features regex --examples &&
     cargo check --bench json &&
     cargo check --bench http &&
-    cargo check --bench mp4 --features mp4
+    cargo check --bench mp4 --features mp4 &&
+    cargo build --no-default-features
